+++ android/app/src/main/java/com/accessmap/MainApplication.java
@@ -1,5 +1,7 @@
 package com.accessmap;
 
+import com.facebook.react.bridge.JSIModulePackage;
+import com.swmansion.reanimated.ReanimatedJSIModulePackage; 
 import android.app.Application;
 import android.content.Context;
 import com.facebook.react.PackageList;
@@ -34,6 +36,11 @@ public class MainApplication extends Application implements ReactApplication {
         @Override
         protected String getJSMainModuleName() {
           return "index";
+        }
+
+        @Override
+        protected JSIModulePackage getJSIModulePackage() {
+          return new ReanimatedJSIModulePackage(); // <- add
         }
       };
