+++ package.json
@@ -13,6 +13,7 @@
     "@react-native-community/masked-view": "^0.1.11",
     "@react-native-community/netinfo": "^6.0.0",
     "@react-native-mapbox-gl/maps": "^8.2.1",
+    "@react-navigation/drawer": "^6.1.3",
     "@react-navigation/native": "^5.9.4",
     "@react-navigation/stack": "^5.14.5",
     "chroma-js": "^2.1.1",
