+++ src/navigation/MapPage.js
@@ -1,5 +1,4 @@
 import React, { Component } from "react";
-import { Drawer } from "native-base";
 import { View, Dimensions, Alert, AccessibilityInfo } from "react-native";
 import NetInfo from "@react-native-community/netinfo";
 import { connect } from "react-redux";
@@ -13,7 +12,6 @@ import LoadingScreen from "../components/LoadingScreen";
 import SpeedLegend from "../components/SpeedLegend";
 import Zooms from "../containers/MapButtons/Zooms";
 import OmniCard from "../containers/OmniCard";
-import SideMenu from "../containers/SideMenu";
 import FeatureCard from "../containers/FeatureCard";
 import RouteBottomCard from "../containers/RouteBottomCard";
 
@@ -32,26 +30,26 @@ class MapPage extends Component {
       screenWidth: Math.round(Dimensions.get("window").width),
     };
     this.onLayout = this.onLayout.bind(this);
-    NetInfo.fetch().then(state => {
+    NetInfo.fetch().then((state) => {
       if (!state.isConnected) {
         Alert.alert(
           "No Internet Connection",
           "Please check your connection to the internet to use AccessMap.",
-          [
-            { text: "OK" }
-          ]
+          [{ text: "OK" }]
         );
       }
     });
   }
 
-  handleFirstConnectivityChange = isConnected => {
+  handleFirstConnectivityChange = (isConnected) => {
     NetInfo.isConnected.removeEventListener(
       "connectionChange",
       this.handleFirstConnectivityChange
     );
 
-    if (isConnected === false) { Alert.alert("You are offline!");}
+    if (isConnected === false) {
+      Alert.alert("You are offline!");
+    }
   };
 
   componentDidUpdate(prevProps) {
@@ -77,71 +75,49 @@ class MapPage extends Component {
     AccessibilityInfo.announceForAccessibility("Showing Map View.");
     return (
       <View style={{ flex: 1 }} onLayout={this.onLayout}>
-        <Drawer
-          // ref={(ref) => {
-          //   this.drawer = ref;
-          // }}
-          style={{
-            drawer: {
-              shadowColor: "#000000",
-              shadowOpacity: 0.8,
-              shadowRadius: 3,
-            },
-            main: { paddingLeft: 3 },
-          }}
-          content={
-            <SideMenu 
-              closeDrawer={() => this.closeDrawer()} 
-              navigation={this.props.navigation}
-            />
-          }
-          onClose={() => this.closeDrawer()}
-        >
-          <View style={Views.page}>
-            <View
-              importantForAccessibility={
-                this.props.openDrawer ? "no-hide-descendants" : "yes"
-              }
-              style={Views.container}
-            >
-              <View style={{ flex: 1 }}> 
-                <View
-                  importantForAccessibility={
-                    this.props.route ? "no-hide-descendants" : "yes"
-                  }
-                  style={{ flex: 1 }}
-                >
-                  <MapView/>
-                </View>
-                {!this.props.viewingDirections &&
-                  !this.props.viewingTripInfo && (
-                    <OmniCard navigation={this.props.navigation} />
-                  )}
-                <Zooms />
+        <View style={Views.page}>
+          <View
+            importantForAccessibility={
+              this.props.openDrawer ? "no-hide-descendants" : "yes"
+            }
+            style={Views.container}
+          >
+            <View style={{ flex: 1 }}>
+              <View
+                importantForAccessibility={
+                  this.props.route ? "no-hide-descendants" : "yes"
+                }
+                style={{ flex: 1 }}
+              >
+                <MapView />
               </View>
-              
-              <SpeedLegend maxIncline={this.props.maxIncline} />
-
-              {this.props.isLoading && <LoadingScreen />}
-              {this.props.pinFeatures && (
-                <FeatureCard navigation={this.props.navigation} />
+              {!this.props.viewingDirections && !this.props.viewingTripInfo && (
+                <OmniCard navigation={this.props.navigation} />
               )}
-              {this.props.route && <RouteBottomCard />}
-              {this.props.viewingDirections && (
-                <Directions
-                  route={this.props.route}
-                  close={() => this.props.closeDirections()}
-                />
-              )}
-              {this.props.viewingTripInfo && (
-                <TripInfo
-                  route={this.props.route}
-                  close={() => this.props.closeTripInfo()}
-                />
-              )}
+              <Zooms />
             </View>
+
+            <SpeedLegend maxIncline={this.props.maxIncline} />
+
+            {this.props.isLoading && <LoadingScreen />}
+            {this.props.pinFeatures && (
+              <FeatureCard navigation={this.props.navigation} />
+            )}
+            {this.props.route && <RouteBottomCard />}
+            {this.props.viewingDirections && (
+              <Directions
+                route={this.props.route}
+                close={() => this.props.closeDirections()}
+              />
+            )}
+            {this.props.viewingTripInfo && (
+              <TripInfo
+                route={this.props.route}
+                close={() => this.props.closeTripInfo()}
+              />
+            )}
           </View>
-        </Drawer>
+        </View>
       </View>
     );
   }
